Conflicts:
	Project Loyal/src/loyal/Battle/Actions/AbilityFactory.java
	Project Loyal/src/loyal/Battle/Actions/AirSlashSingleTargetOffenSiveAbility.java
	Project Loyal/src/loyal/Battle/Actions/BackSlashSingleTargetOffenSiveAbility.java
	Project Loyal/src/loyal/Battle/Actions/DarkOrbOffensiveSingleTargetAbility.java
	Project Loyal/src/loyal/Battle/Actions/DeffensiveAbility.java
	Project Loyal/src/loyal/Battle/Actions/DesperateMoveMultiTargetOffensiveAbility.java
	Project Loyal/src/loyal/Battle/Actions/MultiTargetDeffensiveAbility.java
	Project Loyal/src/loyal/Battle/Actions/SoulArrowOffensiveSingleTarget.java
	Project Loyal/src/loyal/Battle/Actions/SoulShowerMultiTargetOffensiveAbility.java
	Project Loyal/src/loyal/Battle/Actions/TauntSingleTargetDeffenSiveAbility.java
	Project Loyal/src/loyal/Battle/Actions/WizzardAbilityFactory.java
	Project Loyal/src/loyal/Battle/Characters/BadGuys.java
	Project Loyal/src/loyal/Battle/Characters/GoodGuys.java
	Project Loyal/src/loyal/Battle/Characters/Interactable.java
	Project Loyal/src/loyal/Battle/Characters/PlayingCharacter.java
	Project Loyal/src/loyal/entities/CharacterAction.java
	Project Loyal/src/loyal/entities/DefensiveAbility.java
	Project Loyal/src/loyal/entities/DeffensiveAbility.java
	Project Loyal/src/loyal/entities/MagicShieldDefensiveSingleTargetAbility.java
	Project Loyal/src/loyal/entities/MultiTargetDefensiveAbility.java
	Project Loyal/src/loyal/entities/MultiTargetDeffensiveAbility.java
	Project Loyal/src/loyal/entities/State.java

+++ Project Loyal/src/loyal/Battle/Characters/ModifyingItem.java
@@ -7,7 +7,7 @@ public abstract class ModifyingItem implements StateDecorator {
 	// When you inherit from this just set valueChange and changedStat, if you
 	// want to change more stats just add them in and call the
 	protected State innerState;
-	protected int[] modifiedStateArray;
+	protected double[] modifiedStateArray;
 	private String name;
 
 
@@ -18,17 +18,17 @@ public abstract class ModifyingItem implements StateDecorator {
 	}
 
 	@Override
-	public int getStat(Stat statChoice) {
+	public double getStat(Stat statChoice) {
 		return modifiedStateArray[statChoice.index];
 	}
 
 	@Override
-	public void setStat(Stat statChoice, int value) {
+	public void setStat(Stat statChoice, double value) {
 		modifiedStateArray[statChoice.index] = value;
 	}
 
 	@Override
-	public int[] getCurrentStatus() {
+	public double[] getCurrentStatus() {
 		return modifiedStateArray;
 	}
